# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog API",
        "description": "REST API for the Blog application. Provides endpoints for managing posts, categories, tags, comments, and user authentication.",
        "contact": {
            "name": "DigiK Team",
            "email": "admin@digikteam.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "{protocol}://{host}",
            "description": "API Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/categories/{slug}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get a single category by slug",
                "operationId": "015a7b9d140957c1a14c67c5318977e2",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Category slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            }
        },
        "/api/posts/{slug}/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "List approved comments for a post",
                "operationId": "e1eafbfff690534f8a53b7a943720a63",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Post slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Submit a comment on a post",
                "operationId": "314e64ab6194b111d98c0b2949fd0704",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Post slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "author_name",
                                    "author_email",
                                    "content"
                                ],
                                "properties": {
                                    "author_name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "author_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "content": {
                                        "type": "string",
                                        "maxLength": 2000,
                                        "minLength": 3,
                                        "example": "Great post!"
                                    },
                                    "parent_id": {
                                        "description": "Parent comment ID for replies",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comment submitted and awaiting moderation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Comment submitted and awaiting moderation."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post not found or comments disabled"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format: Bearer {your_token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}